sql/migrate_add_columns.py
import psycopg2
from dotenv import load_dotenv
import os

load_dotenv()

conn = psycopg2.connect(
    host=os.getenv('DB_HOST'),
    port=os.getenv('DB_PORT'),
    database=os.getenv('DB_NAME'),
    user=os.getenv('DB_USER'),
    password=os.getenv('DB_PASSWORD')
)

with open('sql/add_columns.sql', 'r') as f:
    sql = f.read()

cur = conn.cursor()
cur.execute(sql)
conn.commit()
cur.close()
conn.close()

print("Columnas agregadas exitosamente!")