requirements.txt
# Orquestación (versiones actualizadas para Python 3.12)
dagster==1.7.16
dagster-webserver==1.7.16

# Data processing
pandas==2.1.3
polars==0.19.13

# Base de datos
psycopg2-binary==2.9.9
sqlalchemy==2.0.23

# API
requests==2.31.0
python-dotenv==1.0.0

# Utils
pyarrow==14.0.1